Exposure Travel Terms
New Subclasses under `travel destination` (GENEPIO_0001783) to be used in the **DataHarmonizer** contextual data *Exposures* vocabulary.

- `travel outside province` (GENEPIO_00001000)
- `travel outside territory` (GENEPIO_00001001)
- `travel outside Canada` (GENEPIO_00001002)

All terms have have the following annotations: `label`, `definition`, and `term editor`. Had `definition source` but since it was the same as the `term editor` I decided it was redundant and did not include it.

*Additional Notes:* Unclear as to why lines `21954` and `22174` jumped to `21991` and `22211` respectively.