GENEPIO_0002173 update
Updated specification axiom from `'specifies value of' some (Country or Municipality)` to `'specifies value of' some (City or Country or Municipality or Province or State or Territory)`. The definition seems appropriate to be inclusive of more options and this way this term can be used more generally by specification `is about` applications (e.g. GENEPIO:0001410, GENEPIO:0001411, GENEPIO:0001412, etc.).

@ddooley let me know if you don't approve these changes.